nodes now store data
nodes now store there own data and the data of the configured at creation nodes to replicate. also changed the save function to save json instead of csv. also storing the nodes it is going to replicate instead of the ones replicating it in the config. next steps will be to share the nodes configuration between all nodes in network.