use graph to store nodes and neighbours. gave nodes coordinates and range.
we now use a directional weighted graph to store the nodes and its neighbours. in the current version the graph is not used for forwarding messages or generating a path. it is currently only used to send data to nodes in range of the current node. next step would be to implement routing.